refactor: global/config 패키지로 이동
