+++ src/main/java/com/backend/connectable/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.backend.connectable;
+package com.backend.connectable.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
